requirements.txt
pycryptodome
requests
mcp>=1.9.0
python-dotenv

# Testing dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
